Power cable work. Added power cables, power grid, cable_registry. Adjusted simulation ticking logic to call Power Producers before Power Grids, then call Power Consumers so that the grids can calculate the demand and other factors.

+++ src/components/PowerProducer.py
@@ -3,7 +3,7 @@ from components.base_component import BaseComponent
 class PowerProducer(BaseComponent):
     def __init__(self, parent, args):
         super().__init__(parent, args)
-        self.watts: float = args["watts"]
+        self.watts: int = args["watts"]
         self.tier: str = args["voltage"]
         self.online = False  # optional (e.g. for fuel machines)
